Move receipt searcher
