+++ src/Stratis.SmartContracts.CLR.Tests/ContractLogHolderTests.cs
@@ -4,6 +4,7 @@ using System.Text;
 using NBitcoin;
 using Stratis.SmartContracts.CLR.ContractLogging;
 using Stratis.SmartContracts.CLR.Serialization;
+using Stratis.SmartContracts.Core;
 using Stratis.SmartContracts.Core.Receipts;
 using Stratis.SmartContracts.Networks;
 using Xunit;
